Remove unneeded hooks
