+++ packages/core/src/data-editor/data-editor.tsx
@@ -47,6 +47,7 @@ import { useColumnSizer } from "./use-column-sizer";
 import { isHotkey } from "../common/is-hotkey";
 import { SelectionBlending, useSelectionBehavior } from "../data-grid/use-selection-behavior";
 import { useCellsForSelection } from "./use-cells-for-selection";
+import { unquote, expandSelection, copyToClipboard } from "./data-editors-fns";
 
 let idCounter = 0;
 
@@ -414,102 +415,16 @@ const DataEditorImpl: React.ForwardRefRenderFunction<DataEditorRef, DataEditorPr
         abortControllerRef.current
     );
 
-    const expandSelection = React.useCallback(
-        (newVal: GridSelection): GridSelection => {
-            const origVal = newVal;
-            if (spanRangeBehavior === "allowPartial" || newVal.current === undefined) return newVal;
-            if (getCellsForSelection !== undefined) {
-                let isFilled = false;
-                do {
-                    if (newVal?.current === undefined) break;
-                    const r: Rectangle = newVal.current?.range;
-                    const cells: (readonly GridCell[])[] = [];
-                    if (r.width > 2) {
-                        const leftCells = getCellsForSelection(
-                            {
-                                x: r.x,
-                                y: r.y,
-                                width: 1,
-                                height: r.height,
-                            },
-                            abortControllerRef.current.signal
-                        );
-
-                        if (typeof leftCells === "function") {
-                            return origVal;
-                        }
-
-                        cells.push(...leftCells);
-
-                        const rightCells = getCellsForSelection(
-                            {
-                                x: r.x + r.width - 1,
-                                y: r.y,
-                                width: 1,
-                                height: r.height,
-                            },
-                            abortControllerRef.current.signal
-                        );
-
-                        if (typeof rightCells === "function") {
-                            return origVal;
-                        }
-
-                        cells.push(...rightCells);
-                    } else {
-                        const rCells = getCellsForSelection(
-                            {
-                                x: r.x,
-                                y: r.y,
-                                width: r.width,
-                                height: r.height,
-                            },
-                            abortControllerRef.current.signal
-                        );
-                        if (typeof rCells === "function") {
-                            return origVal;
-                        }
-                        cells.push(...rCells);
-                    }
-
-                    let left = r.x - rowMarkerOffset;
-                    let right = r.x + r.width - 1 - rowMarkerOffset;
-                    cells.forEach(row =>
-                        row.forEach(cell => {
-                            if (cell.span === undefined) return;
-                            left = Math.min(cell.span[0], left);
-                            right = Math.max(cell.span[1], right);
-                        })
-                    );
-                    if (left === r.x - rowMarkerOffset && right === r.x + r.width - 1 - rowMarkerOffset) {
-                        isFilled = true;
-                    } else {
-                        newVal = {
-                            current: {
-                                cell: newVal.current.cell ?? [0, 0],
-                                range: {
-                                    x: left + rowMarkerOffset,
-                                    y: r.y,
-                                    width: right - left + 1,
-                                    height: r.height,
-                                },
-                                rangeStack: newVal.current.rangeStack,
-                            },
-                            columns: newVal.columns,
-                            rows: newVal.rows,
-                        };
-                    }
-                } while (!isFilled);
-            }
-            return newVal;
-        },
-        [getCellsForSelection, rowMarkerOffset, spanRangeBehavior]
-    );
-
     const setGridSelection = React.useCallback(
         (newVal: GridSelection, expand: boolean): void => {
             if (expand) {
-                newVal = expandSelection(newVal);
+                newVal = expandSelection(
+                    newVal,
+                    getCellsForSelection,
+                    rowMarkerOffset,
+                    spanRangeBehavior,
+                    abortControllerRef.current
+                );
             }
             if (onGridSelectionChange !== undefined) {
                 onGridSelectionChange(shiftSelection(newVal, -rowMarkerOffset));
@@ -517,7 +432,7 @@ const DataEditorImpl: React.ForwardRefRenderFunction<DataEditorRef, DataEditorPr
                 setGridSelectionInner(newVal);
             }
         },
-        [onGridSelectionChange, rowMarkerOffset, expandSelection]
+        [onGridSelectionChange, getCellsForSelection, rowMarkerOffset, spanRangeBehavior]
     );
 
     const onDelete = React.useCallback<NonNullable<DataEditorProps["onDelete"]>>(
@@ -1545,50 +1460,6 @@ const DataEditorImpl: React.ForwardRefRenderFunction<DataEditorRef, DataEditorPr
         ]
     );
 
-    const copyToClipboard = React.useCallback(
-        (cells: readonly (readonly GridCell[])[], columnIndexes: readonly number[]) => {
-            function escape(str: string): string {
-                if (/\n|"|\t/.test(str)) {
-                    str = `"${str.replace(/"/g, `""`)}"`;
-                }
-                return str;
-            }
-
-            const formatCell = (cell: GridCell, index: number): string => {
-                const colIndex = columnIndexes[index];
-                if (cell.span !== undefined && cell.span[0] !== colIndex) return "";
-                switch (cell.kind) {
-                    case GridCellKind.Text:
-                    case GridCellKind.Number:
-                        return escape(cell.displayData);
-                    case GridCellKind.Markdown:
-                    case GridCellKind.RowID:
-                    case GridCellKind.Uri:
-                        return escape(cell.data);
-                    case GridCellKind.Image:
-                    case GridCellKind.Bubble:
-                        return cell.data.reduce((pv, cv) => `${escape(pv)},${escape(cv)}`);
-                    case GridCellKind.Boolean:
-                        return cell.data ? "TRUE" : "FALSE";
-                    case GridCellKind.Loading:
-                        return "#LOADING";
-                    case GridCellKind.Protected:
-                        return "************";
-                    case GridCellKind.Drilldown:
-                        return cell.data.map(i => i.text).reduce((pv, cv) => `${escape(pv)},${escape(cv)}`);
-                    case GridCellKind.Custom:
-                        return escape(cell.copyData);
-                    default:
-                        assertNever(cell);
-                }
-            };
-
-            const str = cells.map(row => row.map(formatCell).join("\t")).join("\n");
-            void window.navigator.clipboard.writeText(str);
-        },
-        []
-    );
-
     // 1 === move one
     // 2 === move to end
     const adjustSelection = React.useCallback(
@@ -2181,74 +2052,6 @@ const DataEditorImpl: React.ForwardRefRenderFunction<DataEditorRef, DataEditorPr
         ]
     );
 
-    function unquote(str: string): string[][] {
-        function descape(s: string): string {
-            if (s.startsWith('"') && s.endsWith('"')) {
-                s = s.slice(1, -1).replace(/""/g, '"');
-            }
-            return s;
-        }
-
-        const enum State {
-            None,
-            inString,
-            inStringPostQuote,
-        }
-
-        const result: string[][] = [];
-        let current: string[] = [];
-
-        let start = 0;
-        let state = State.None;
-        str = str.trim().replace(/\r\n/g, "\n");
-        let index = 0;
-        for (const char of str) {
-            switch (state) {
-                case State.None:
-                    if (char === "\t" || char === "\n") {
-                        current.push(str.slice(start, index));
-                        start = index + 1;
-
-                        if (char === "\n") {
-                            result.push(current);
-                            current = [];
-                        }
-                    } else if (char === `"`) {
-                        state = State.inString;
-                    }
-                    break;
-                case State.inString:
-                    if (char === `"`) {
-                        state = State.inStringPostQuote;
-                    }
-                    break;
-                case State.inStringPostQuote:
-                    if (char === '"') {
-                        state = State.inString;
-                    } else if (char === "\t" || char === "\n") {
-                        current.push(descape(str.slice(start, index)));
-                        start = index + 1;
-
-                        if (char === "\n") {
-                            result.push(current);
-                            current = [];
-                        }
-                        state = State.None;
-                    } else {
-                        state = State.None;
-                    }
-                    break;
-            }
-
-            index++;
-        }
-        if (start < str.length) {
-            current.push(descape(str.slice(start, str.length)));
-        }
-        result.push(current);
-        return result;
-    }
-
     const onPasteInternal = React.useCallback(async () => {
         if (!keybindings.paste) return;
         function pasteToCell(inner: InnerGridCell, target: Item, toPaste: string): boolean {
@@ -2403,15 +2206,7 @@ const DataEditorImpl: React.ForwardRefRenderFunction<DataEditorRef, DataEditorPr
                 copyToClipboard(toCopy, cols);
             }
         }
-    }, [
-        columnsIn.length,
-        copyToClipboard,
-        getCellsForSelection,
-        gridSelection,
-        keybindings.copy,
-        rowMarkerOffset,
-        rows,
-    ]);
+    }, [columnsIn.length, getCellsForSelection, gridSelection, keybindings.copy, rowMarkerOffset, rows]);
 
     useEventListener("copy", onCopy, window, true, false);
