+++ packages/core/src/data-editor/stories/data-editor-beautiful.stories.tsx
@@ -232,7 +232,6 @@ export const AddData: React.VFC = () => {
                 getCellsForSelection={getCellsForSelection}
                 rowMarkers={"both"}
                 onPaste={true}
-                fillHandle={true}
                 onCellEdited={setCellValue}
                 trailingRowOptions={{
                     sticky: true,
@@ -246,6 +245,61 @@ export const AddData: React.VFC = () => {
     );
 };
 (AddData as any).parameters = {
+    options: {
+        showPanel: false,
+    },
+};
+
+export const FillHandle: React.VFC = () => {
+    const { cols, getCellContent, setCellValueRaw, setCellValue, getCellsForSelection } = useMockDataGenerator(
+        60,
+        false
+    );
+
+    const [numRows, setNumRows] = React.useState(50);
+
+    const onRowAppended = React.useCallback(() => {
+        const newRow = numRows;
+        for (let c = 0; c < 6; c++) {
+            const cell = getCellContent([c, newRow]);
+            setCellValueRaw([c, newRow], clearCell(cell));
+        }
+        setNumRows(cv => cv + 1);
+    }, [getCellContent, numRows, setCellValueRaw]);
+
+    return (
+        <BeautifulWrapper
+            title="Add data"
+            description={
+                <>
+                    <Description>Fill handles can be used to downfill data with the mouse.</Description>
+                    <MoreInfo>
+                        Just click and drag, the top row will be copied down. Enable using the{" "}
+                        <PropName>fillHandle</PropName> prop.
+                    </MoreInfo>
+                </>
+            }>
+            <DataEditor
+                {...defaultProps}
+                getCellContent={getCellContent}
+                columns={cols}
+                getCellsForSelection={getCellsForSelection}
+                rowMarkers={"both"}
+                onPaste={true}
+                fillHandle={true}
+                onCellEdited={setCellValue}
+                trailingRowOptions={{
+                    sticky: true,
+                    tint: true,
+                    hint: "New row...",
+                }}
+                rows={numRows}
+                onRowAppended={onRowAppended}
+            />
+        </BeautifulWrapper>
+    );
+};
+(FillHandle as any).parameters = {
     options: {
         showPanel: false,
     },
